refactor(bpa-server): reorganize static_routes into agent.rs and loader.rs
